Refactor for MIPS build with isolated Cargo.lock
- Move CLI logic from main.rs to new src/cli.rs module
- Simplify main.rs to 3-line wrapper calling gateway_rs::cli::main()
- Feature-gate http/http-serde deps (modern vs legacy versions)
- Add unified http re-exports in proto.rs for version compatibility
- Update all source files to use crate::proto::http instead of http directly
- Simplify mips-build/Cargo.toml to depend on gateway-rs with only
  3 version pins for conflicting transitive deps (ed25519, signature)
- Fix Makefile.toml: add clear=true to prevent cargo-make default merge
- Pin shared deps (clap, config, time, thiserror) for Rust 1.72.1 compat

The MIPS build requires a separate Cargo.lock because helium-crypto 0.8.3
and 0.9.4 have conflicting transitive dependencies (zeroize, signature).
