+++ mips-build/src/main.rs
@@ -1,5 +1,3 @@
-// This is a stub file for lockfile generation.
-// The actual gateway code is built from the root using this lockfile.
-fn main() {
-    println!("MIPS lockfile generator stub - not meant to be run");
+fn main() -> ! {
+    gateway_rs::cli::main()
 }
