+++ src/keypair.rs
@@ -3,9 +3,9 @@ use crate::proto::crypto::ecc608;
 #[cfg(feature = "tpm")]
 use crate::proto::crypto::tpm;
 use crate::proto::crypto::{KeyTag, KeyType, Network};
+use crate::proto::http::Uri;
 use crate::{DecodeError, Error, Result};
 use async_trait::async_trait;
-use http::Uri;
 use rand::rngs::OsRng;
 use serde::{de, Deserializer};
 #[cfg(feature = "ecc608")]
